# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UOU Infinite - Institutional Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: courseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}/credential:
    get:
      operationId: getStudentCredential
      tags: [students]
      summary: Get student QR credential data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Credential data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentCredential"
...
  /lecturers:
    get:
      operationId: listLecturers
      tags: [lecturers]
      summary: List all lecturers
      responses:
        "200":
          description: List of lecturers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lecturer"
    post:
      operationId: createLecturer
      tags: [lecturers]
      summary: Create a lecturer profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLecturerBody"
      responses:
        "201":
          description: Created lecturer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecturer"
  /lecturers/{id}:
    get:
      operationId: getLecturer
      tags: [lecturers]
      summary: Get lecturer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lecturer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecturer"
    patch:
      operationId: updateLecturer
      tags: [lecturers]
      summary: Update lecturer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLecturerBody"
      responses:
        "200":
          description: Updated lecturer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecturer"

  # Admissions
...
  /enrollments:
    get:
      operationId: listEnrollments
      tags: [enrollments]
      summary: List enrollments
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
        - name: courseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
    post:
      operationId: createEnrollment
      tags: [enrollments]
      summary: Enroll student in a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnrollmentBody"
      responses:
        "201":
          description: Created enrollment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /enrollments/{id}:
    delete:
      operationId: deleteEnrollment
      tags: [enrollments]
      summary: Remove enrollment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Dashboard / Founder War Room
...
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /openai/conversations/{id}/messages:
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a message and receive a streaming response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
components:
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        studentId:
          type: string
        department:
          type: string
        level:
          type: string
        status:
          type: string
        gpa:
          type: ["number", "null"]
        lastLoginAt:
          type: ["string", "null"]
          format: date-time
        loginCount:
          type: integer
        credentialToken:
          type: string
        skills:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - email
        - studentId
        - department
        - level
        - status
        - loginCount
        - credentialToken
        - skills
        - createdAt
    UpdateStudentBody:
      type: object
      properties:
        department:
          type: string
        level:
          type: string
        status:
          type: string
        skills:
          type: string
    StudentCredential:
      type: object
      properties:
        studentId:
          type: string
        name:
          type: string
        department:
          type: string
        skills:
          type: string
        credentialToken:
          type: string
        issuedAt:
          type: string
          format: date-time
        qrData:
          type: string
      required:
        - studentId
        - name
        - department
        - skills
        - credentialToken
        - issuedAt
        - qrData
...
    Lecturer:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        department:
          type: string
        specialization:
          type: string
        courseCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - email
        - department
        - specialization
        - courseCount
        - createdAt
    CreateLecturerBody:
      type: object
      properties:
        userId:
          type: integer
        department:
          type: string
        specialization:
          type: string
      required:
        - userId
        - department
        - specialization
    UpdateLecturerBody:
      type: object
      properties:
        department:
          type: string
        specialization:
          type: string

    # Admissions
...
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        courseId:
          type: integer
        courseName:
          type: string
        studentName:
          type: string
        grade:
          type: ["string", "null"]
        enrolledAt:
          type: string
          format: date-time
      required:
        - id
        - studentId
        - courseId
        - courseName
        - studentName
        - enrolledAt
    CreateEnrollmentBody:
      type: object
      properties:
        studentId:
          type: integer
        courseId:
          type: integer
      required:
        - studentId
        - courseId

    # Dashboard
...
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages