# lib/api-zod/src/generated/types/studentCredential.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UOU Infinite - Institutional Operating System API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentCredential {
  studentId: string;
  name: string;
  department: string;
  skills: string;
  credentialToken: string;
  issuedAt: Date;
  qrData: string;
}