# lib/api-zod/src/generated/types/updateStudentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UOU Infinite - Institutional Operating System API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateStudentBody {
  department?: string;
  level?: string;
  status?: string;
  skills?: string;
}